requirements.txt
gym==0.17.2
mujoco-py==2.0.2.0
-e git+https://github.com/martius-lab/smart-settings@85e42b2eff23c3ea1ddc544e8a1784841da64735#egg=smart_settings
numpy==1.23.5
tqdm==4.64.1
forwardable==0.4.1
scikit-learn==1.2.0
scipy==1.9.3
matplotlib==3.6.2
pre-commit==2.20.0
black==22.12.0
h5py==3.7.0
GitPython==3.1.29
tensorboardX==2.5.1
tensorboard==2.11.0
jupyter==1.0.0
imageio==2.22.4
imageio-ffmpeg==0.4.7
seaborn==0.12.1